requirements.txt
awacs==0.6.2
boto3==1.4.4
botocore==1.5.76
cached-property==1.3.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
click-log==0.1.8
docutils==0.13.1
Flask==0.12.2
Flask-SQLAlchemy==2.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
MarkupSafe==1.0
marshmallow==2.13.5
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-box==3.0.1
python-dateutil==2.6.0
requests==2.18.1
ruamel.yaml==0.15.15
s3transfer==0.1.10
six==1.10.0
SQLAlchemy==1.1.11
troposphere==1.9.4
urllib3==1.21.1
Werkzeug==0.12.2
